The Desired Columns
Took a while before I could figure out which column links to which and
what columns I need for an overview of units. I will leave upgrades and
battle gears for now, but I have also included the data needed to
achieve this. And I just realized I did not build the project.