No Training Time For Now
Abandoned training time of units from buildings because of a lot of
missing units and my ineffective methods to calculate for total training
time. Also decided to publish only a simple unit and weapon data for
quick review, but not yet ready to publish because I need to clan names,
unit, names, damage class for melee and missile weapons.